fix(analysers): rename GraphQL analyser and add dependency
