+++ pyproject.toml
@@ -13,6 +13,7 @@ dependencies = [
   "tomli>=1.0.0",
   "jinja2>=3.1",
   "pyyaml>=6.0",
+  "graphql-core>=3.2",
 ]
 
 classifiers = [
